+++ package.json
@@ -43,7 +43,6 @@
   "devDependencies": {
     "@parcel/optimizer-data-url": "^2.8.0",
     "@parcel/transformer-inline-string": "^2.8.0",
-    "@parcel/transformer-sass": "^2.8.0",
     "@types/halfmoon": "^1.0.2",
     "@types/node": "^18.11.9",
     "@types/react": "^18.0.26",
