+++ src/webview/index.html
@@ -10,12 +10,12 @@
       href="data:image/svg+xml,<svg xmlns=%22http://www.w3.org/2000/svg%22 viewBox=%220 0 100 100%22><text y=%22.9em%22 font-size=%2290%22>⏲️</text></svg>"
     />
     <style>
-      @import "./index.scss";
+      @import "./styles/index.css";
     </style>
   </head>
   <body class="with-custom-webkit-scrollbars with-custom-css-scrollbars" data-set-preferred-mode-onload="true">
     <script type="module">
-      import "./index.tsx";
+      import "./scripts/index.tsx";
     </script>
   </body>
 </html>
