+++ src/webview/scripts/components/Root/StickyAlertsContainer/StickyAlertsContainer.tsx
@@ -1,0 +1,7 @@
+import { useEffect } from "react";
+import halfmoon from "halfmoon";
+
+export function StickyAlertsContainer() {
+  useEffect(() => halfmoon.onDOMContentLoaded(), []);
+  return <div className="sticky-alerts"></div>;
+}
