+++ src/webview/scripts/enumerations.ts
@@ -1,0 +1,5 @@
+export enum CurrentScreen {
+  InitialScreen,
+  JoinScreen,
+  TimerScreen,
+}
